HACK libmysql/database_msql.cpp to compile
This might break the code itself, but at least it compiles and
cleaning up the buildsystem is all I am interested in right now.

The code looks so unnecessarily repetitive that I am not going to
touch it anyway for the required refactoring without knowing what
it is actually supposed to do.

This is the first commit for updating the automake buildsystem.