Make "make distcheck" run
This mainly involves adding a bunch of source files and image files
to the appropriate _SOURCES or EXTRA_DIST variable, but also adds
source files with translations to POTFILES.in while removing the
non-existing file src/main.c from POTFILES.in.

While dist tarballs as a distribution method might not be the
first priority, "make distcheck" still makes sure that things
like clean, uninstall, etc. work properly, so this is worthwhile.